[Feat] MockStationLIst update
