+++ Projects/Domain/Sources/RepositoryInterface/StationListRepository.swift
@@ -1,5 +1,5 @@
 //
-//  SearchRepository.swift
+//  StationListRepository.swift
 //  Domain
 //
 //  Created by 유하은 on 2024/02/27.
@@ -10,6 +10,6 @@ import Foundation
 
 import RxSwift
 
-public protocol SearchRepository {
+public protocol StationListRepository {
     var searchResponse: BehaviorSubject<[BusStopInfoResponse]> { get }
 }
