+++ Projects/Feature/SearchFeature/Demo/AppDelegate.swift
@@ -2,7 +2,6 @@ import UIKit
 
 import Core
 import Domain
-import HomeFeature
 import FeatureDependency
 
 import RxSwift
@@ -10,6 +9,7 @@ import RxSwift
 @main
 final class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        register()
         return true
     }
     
@@ -20,5 +20,12 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
     }
     
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
+    }
+    
+    func register() {
+//        DIContainer.register(
+//            type: SearchRepository.self,
+//            )
+//        )
     }
 }
