+++ Projects/Feature/SearchFeature/Sources/ViewController/SearchViewController.swift
@@ -129,7 +129,6 @@ public final class SearchViewController: UIViewController, UITableViewDelegate {
     
     public override func viewDidLoad() {
         super.viewDidLoad()
-        viewModel.useCase.loadBusStopInfoList()
         configureDataSource()
         configureUI()
         bind()
